Added boring entities. Can create and render multi-subtextures. Added sprite system. FOR SOME REASON, can only render one sprite a time. LIKE HOW??

+++ Gonk/src/Gonk/Debug/Instrumentor.h
@@ -27,7 +27,7 @@ struct ProfileResult
 {
     std::string Name;
     long long Start, End;
-    uint32_t ThreadID;
+    size_t ThreadID;
 };
 
 struct InstrumentationSession
